+++ driver-observe/src/main/java/fr/ird/driver/observe/dao/data/ps/localmarket/BatchDao.java
@@ -42,7 +42,7 @@ public class BatchDao extends AbstractDataDao<Batch> {
         super.fill(result, rs);
         result.setDate(rs.getDate(6));
         result.setCount(rs.getInt(7));
-        result.setWeight(rs.getFloat(8));
+        result.setWeight(rs.getDouble(8));
         result.setWeightComputedSource(rs.getString(9));
         result.setOrigin(rs.getString(10));
         result.setComment(rs.getString(11));
